feat(ci): move 21 mocked coordinator tests to unit suite, shrink acceptance-gate job
21 deterministic tests in test/coordinator/acceptance-gate.test.ts used
vi.fn().mockImplementation() to mock instrumentWithRetry and ran in milliseconds
despite living in the LLM acceptance-gate CI job (now 120-minute cap). Moving them
to the regular unit test suite (npm test) eliminates their unnecessary presence from
the acceptance-gate timing budget.

New file: test/coordinator/coordinator.test.ts (21 tests, no API calls)
- Phase 5 SCH Tier 2 Checks (6)
- SCH-001/002 unconditionally blocking (4)
- PRD 31 Per-File Schema Extension Writing (5)
- Phase 5 Checkpoint and Drift Integration (3)
- End-of-run failure handling (2)
- PRD 700 Dependency-aware file ordering (1)

Remaining in acceptance-gate.test.ts (11 tests):
- Phase 4 Coordinator (5 LLM-calling)
- Phase 5 Schema Integration (5 LLM-calling)
- runLiveCheck M5 end-to-end SDK injection (1 real integration test)

Two test assertions corrected during the move:
- flagCtx.testOutput → flagCtx.failureMessage (current EndOfRunFlagContext field name)
- Scenario B directError → testOutput (what isDirectError() actually reads)
