+++ PROGRESS.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
+- (2026-05-08) Moved 21 deterministic mocked tests out of the acceptance-gate CI job into the regular unit test suite (issue #835). The coordinator acceptance gate file contained 33 tests but only 10 required real Anthropic API calls; the remaining 23 used `vi.fn().mockImplementation()` for `instrumentWithRetry` and ran in milliseconds. Moved blocks: Phase 5 SCH Tier 2 Checks (6 tests), SCH-001/002 unconditionally blocking (4 tests), PRD 31 Per-File Schema Extension Writing (5 tests), Phase 5 Checkpoint and Drift Integration (3 tests), End-of-run failure handling (2 tests), PRD 700 Dependency-aware file ordering (1 test). These now live in `test/coordinator/coordinator.test.ts` and run under `npm test`, freeing the coordinator acceptance-gate job to contain only its 10 LLM-calling tests and the runLiveCheck M5 integration test. Two test assertions were corrected in the process: `flagCtx.testOutput` renamed to `flagCtx.failureMessage` (matching the current `EndOfRunFlagContext` type), and Scenario B's `directError` field renamed to `testOutput` (matching what `isDirectError()` actually reads).
 - (2026-05-08) Culled three redundant tests from the commit-story-v2 acceptance gate suite (`index.js`, `summary-manager.js`, `summary-detector.js`) and raised the CI job timeout from 80→130 minutes (step: 75→120 minutes). The three removed tests covered the same patterns as remaining tests: `index.js` duplicated the CLI-entry-point + schema-extension pattern of `summarize.js`; `summary-manager.js` and `summary-detector.js` duplicated the async-file-operations + graceful-degradation-catch pattern of `journal-manager.js`. The suite consistently timed out at 75 minutes with 8 LLM-calling tests; reducing to 5 and doubling the timeout provides headroom for the remaining tests. The coordinator acceptance gate job also benefits from the extended timeout.
 
 ### Fixed
